refactor: split rolling period results
