+++ quality/baseline_report.md
@@ -13,11 +13,11 @@ completion claim.
 ## Generation Identity
 
 - Git branch: `feat/enterprise-risk-refactor-continuation`
-- Git commit: `62e70779203ac2bd1cb8b5dccba6abe923bae028`
+- Git commit: `d03743fbbc102616a92febb5cfa937bbdec87063`
 
 ## Current Code Size
 
-- Python source files under `src/`: 214
+- Python source files under `src/`: 215
 - Python test files under `tests/`: 103
 - Python packages under `src/`: 9
 - API entry point route/middleware/handler decorators in `src/app/main.py`: 0
@@ -26,7 +26,6 @@ completion claim.
 
 | Path | Lines | Bytes |
 | --- | --- | --- |
-| src/app/services/rolling_engine.py | 310 | 10403 |
 | src/app/services/rolling_stateful_inputs.py | 302 | 10592 |
 | src/app/services/risk/calculation_orchestrator.py | 301 | 9984 |
 | src/app/services/risk/helpers.py | 297 | 10019 |
@@ -46,6 +45,7 @@ completion claim.
 | src/app/services/concentration/response_builder.py | 244 | 9409 |
 | src/app/services/attribution_engine.py | 244 | 7513 |
 | src/app/services/rolling_risk_free_dependency.py | 242 | 7826 |
+| src/app/services/concentration/parsing.py | 231 | 7793 |
 
 ### Largest Functions And Classes
 
@@ -96,12 +96,12 @@ All checks passed!
 - Ruff format check: passed
 
 ```text
-332 files already formatted
+333 files already formatted
 ```
 - Type checking: passed
 
 ```text
-Success: no issues found in 317 source files
+Success: no issues found in 318 source files
 ```
 - Unit coverage snapshot: passed
 
@@ -115,10 +115,10 @@ Success: no issues found in 317 source files
 ...
 src\app\services\rolling_stateful_inputs.py                   80      1     18      1    98%   102
 ------------------------------------------------------------------------------------------------------
-TOTAL                                                       6061     91   1028     73    98%
+TOTAL                                                       6067     91   1028     73    98%
 
-171 files skipped due to complete coverage.
-550 passed in 8.53s
+172 files skipped due to complete coverage.
+550 passed in 8.60s
 ```
 
 ## Complexity And Maintainability Snapshot
@@ -148,7 +148,7 @@ src\app\services\risk\__init__.py - A (100.00)
 - Dependency hygiene: passed
 
 ```text
-Scanning 214 files...
+Scanning 215 files...
 
 Success! No dependency issues found.
 ```
@@ -247,7 +247,7 @@ tests/unit/test_upstream_errors.py::test_classify_upstream_transport_error_matri
 tests/unit/test_upstream_errors.py::test_classify_upstream_transport_error_matrix[exc1-503-UPSTREAM_UNAVAILABLE-transport]
 tests/unit/test_upstream_errors.py::test_invalid_payload_and_missing_data_carry_structured_categories
 
-550 tests collected in 1.61s
+550 tests collected in 1.57s
 ```
 - Import-linter report-only: passed
